requirements.txt
openai>=1.0.0
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0
typer>=0.9.0
pydantic>=2.0.0
colorama>=0.4.6 